Fixing issue where viewUserPermission couldn't find an email
